refactor software consumers
